+++ BlazorTutorial/EmployeeManagement.Web/Pages/EditEmployeeBase.cs
@@ -90,11 +90,21 @@ namespace EmployeeManagement.Web.Pages
                 NavigationManager.NavigateTo("/");
             }
         }
+        protected PragimTech.Components.ConfirmBase DeleteConfirmation { get; set; }
 
-        protected async Task Delete_Click() 
+        protected void Delete_Click()
         {
-            await EmployeeService.DeleteEmployee(Employee.EmployeeId);
-            NavigationManager.NavigateTo("/");
+            DeleteConfirmation.Show();
+        }
+
+        protected async Task ConfirmDelete_Click(bool deleteConfirmed)
+        {
+            if (deleteConfirmed)
+            {
+                await EmployeeService.DeleteEmployee(Employee.EmployeeId);
+                NavigationManager.NavigateTo("/");
+
+            }
         }
     }
 }
